Fill context on login with third party & refactor
